pyproject.toml
[tool.poetry]
name = "github_telegram_bot"
version = "0.1.0"
description = "Github webhooks handles for Telegram"
authors = ["Martin Kagamino Lehoux <martin@lehoux.net>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.0"
dacite = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
flake8 = "^3.8.4"
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
